requirements.txt
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
db-dtypes==1.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
idna==3.10
numpy==2.0.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pandas-gbq==0.26.1
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
